+++ src/commands/mod.rs
@@ -1,3 +1,4 @@
+pub mod agps_date;
 pub mod download_agps;
 pub mod download_tracks;
 pub mod get_settings;